requirements/base.txt
tornado==6.0.2
sqlalchemy==1.2.15
ujson==1.35
redis==3.0.1
fire==0.1.3
alembic==1.0.5
#lxml==4.2.5
pymysql==0.9.3
looter==2.11
markdown==3.0.1
imgkit==1.0.1
rsa==4.0
python-dotenv==0.10.1
requests>=2.22.0
#feedparser>=5.2.1
#apscheduler>=3.6.0
